[func] Keep translations when deleting attachment by CLI

+++ Services/AttachmentsManager.php
@@ -11,22 +11,26 @@ class AttachmentsManager
 {
     public function deleteAttachment($attachment_id)
     {
-        remove_action('delete_attachment', [$this, 'deleteAttachment']);
+        if(defined('WP_CLI')) {
+            output_success(sprintf('WP_CLI is defined. Keep attachment %s translations', $attachment_id));
+        } else {
+            remove_action('delete_attachment', [$this, 'deleteAttachment']);
 
-        $deleted_attachement = wp_cache_get('woody_deleted_attachement', 'woody');
-        if (empty($deleted_attachement)) {
-            $deleted_attachement = [];
-        }
+            $deleted_attachement = wp_cache_get('woody_deleted_attachement', 'woody');
+            if (empty($deleted_attachement)) {
+                $deleted_attachement = [];
+            }
 
-        if (wp_attachment_is_image($attachment_id) && is_array($deleted_attachement) && !in_array($attachment_id, $deleted_attachement)) {
-            // Remove translations
-            $translations = pll_get_post_translations($attachment_id);
-            $deleted_attachement = array_merge($deleted_attachement, array_values($translations));
-            wp_cache_set('woody_deleted_attachement', $deleted_attachement, 'woody');
+            if (wp_attachment_is_image($attachment_id) && is_array($deleted_attachement) && !in_array($attachment_id, $deleted_attachement)) {
+                // Remove translations
+                $translations = pll_get_post_translations($attachment_id);
+                $deleted_attachement = array_merge($deleted_attachement, array_values($translations));
+                wp_cache_set('woody_deleted_attachement', $deleted_attachement, 'woody');
 
-            foreach ($translations as $t_attachment_id) {
-                if ($t_attachment_id != $attachment_id) {
-                    wp_delete_attachment($t_attachment_id);
+                foreach ($translations as $t_attachment_id) {
+                    if ($t_attachment_id != $attachment_id) {
+                        wp_delete_attachment($t_attachment_id);
+                    }
                 }
             }
         }
